[ETHFW-276] ethfw: Add KWStyle script
Add initial vesion of the KWStyle check script. The KWStyle rules are
for CPSW LLD code only (i.e. due to variable and function prefixes).

To run the script:

 $ cd <pdk>/packages/ti/drv/cpsw
 $ <ethfw>/internal_docs/scripts/dev/cpsw_kwstyle.bash

The KWStyle errors will be logged in a temporary file whose name is
printed at the end of the script execution.

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>