app_ndk_switch: Use MAC_PORT_1 as primary port for NIMU/NDK
Temporarily use MAC port 1 as the primary port for NIMU/NDK. This port is
the one being used as the host port address. This change requires CPSW NIMU
implementation to have the same port number.

The main reason for this change is the EVM-related CRC problems observed
in J7 GESI board. This change is to be reverted when EVM problem is fixed
or NIMU dependency on specific MAC port is addressed.

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>