ethfw: build: Fix parallel build
Parallel execution of ethfw build was breaking due to missing dependency
of apps on PDK, causing apps to fail to compile due to missing symbols
of PDK libraries not yet build.

concerto's 'all' target depends on 'depend' target. So using 'depend' to
convey the general dependency on PDK seemed appropriate.

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>