build: Allow building multiple SOCs with 'ethfw_all' make rule
Add support for building EthFw binaries for all supported SoCs in a single
make call.

SoCs to build for is controlled via BUILD_SOC_LIST variable, which by default
is set to J721E and J7200. If user wants to build EthFw for a single SoC,
it should do so as follows:

Build EthFw for J721E:
$ make -s -j ethfw_all BUILD_SOC_LIST:=j721e

Build EthFw for J7200:
$ make -s -j ethfw_all BUILD_SOC_LIST:=j7200

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>