ETHFW-1273 - apps: j7200: Update IPC vring memory
IPC Vring memory address has been changed for j7200 SoC to 0xA4000000.
The address for j721e remains at 0xAA000000.

The intention of this address change is to remove the gap that would
otherwise exist in j7200 between firmware locations in memory and Vring
memory.

Note that this address change is required by the IPC driver.

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>