ethfw: Use port count-agnostic CPSW types
The CPSW type parameter passed in ATTACH and ATTACH_EXT commands was defined
in terms of the number of ports, i.e. CPSW2G, CPSW5G, CPSW9G.

For their virtual MAC interface, the remote clients don't really need to know
how many ports the CSPW switch has.  So, we are moving to a CPSW type
definition which is agnostic of the port count and instead defines the types
as MCU or MAIN depending on the domain the belong.

This simplifies remote clients code (i.e. J721E or J7200) as they can just
pass MAIN_CPSW type to use the virtual MAC interface from Ethernet Firmware.

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>