ETHFW-1671 - ethremotecfg: Don't assert if RPMessage_send() fails for disabled core
EthFw now enables server-to-client notify service for MPU1_0 and MCU2_1
cores. But when one core is not enabled, the RPMessage_send() call to
that core will fail, and that's an expected condition under those
condition and assertion should not occur.

The assertion fixed by this change is:
Assertion @ Line: 2375 in ethremotecfg/server/src/cpsw_proxy_server.c: IPC_SOK == rtnVal : failed !!

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>