ETHFW-1672 - client: Fix unaligned access notifyService message buffer
A data abort was being observed on the EthFw TI-RTOS client application
when a notification message was received and the register callback was
called.

The abort was reported by BIOS as a "bad calling context" failure
pointing to a semaphore not being called from a Task.  This is not
accurate as the problem was actually caused by an unaligned access in a
"ldm" ARM instruction resulting from accessing hwPushNum and/or
timeStamp members of CpswRemoteNotifyService_HwPushMsg struct.

When issue was observed, stack variable of type
CpswRemoteNotifyService_HwPushMsg was at a 16-bit aligned address, not
32-bit aligned which is required for "ldm" instruction.

This issue was recently introduced by "033dff7 [PSDKQA-187][ETHFW-1645]
Update PTP Support for Remote Client", so the fix consist of reverting
the allocation of IPC message buffer back to a 64-bit array instead of
8-bit array.

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>