ETHFW-1692 - apps: server: Flush IPC trace buffer
The IPC trace buffer may be in a cached memory, so the buffer has to be
flushed periodically to ensure that latest logs are observed on remote
cores (i.e. Linux).

The IPC trace buffer flush mechanism implemented here is based on IPC
driver example app.  The flush operation is hooked as a BIOS idle
function so it runs when system is not doing any important task.

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>