server: Remove netif link callback
The netif link callback reports changes in the PHY link status. The PHY
whose status is to be reported is straighforward when operating in MAC
mode, but when operating in Switch mode, the link status cannot be that
of any PHY connected to the switch, instead, it should be the status of
the link between CPSW host port and Switch R5F core, which is always
linked.

Instead, EthFw reports link up when at least one PHY is linked to a
partner device, and link down after last partner device disconnects.
This can be misleading as user might expect to see link up / link down
prints whenever a partner device is connected/disconnected.

Due to above reason, the link status callback will not be used.

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>