ethremotecfg: Port definition using 'virtual port'
Move away from a remote core configuration based on strings to a
mechanism based on 'virtual ports'.

At this point, 'virtual port' refers to a port in the CPSW switch that
is accessed via DMA and behaves as another port in the switch (with some
known limitations about forwarding among virtual ports).

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>