ethremotecfg: proxy: Add CpswProxy_connect()
Split CpswProxy_init() into two operations: remote_device init and
connection with server.

The connection to the server may take long time if server is not
available and potentially not happen ever in case of server not
loaded or any fatal error. If that happens, client application would
remain blocked, which is not a valid condition.

The expected call sequence for client now is:
 CpswProxy_init() - once per core
 CpswProxy_connect() - call as many times as needed until connection
                       is established
 CpswProxy_open() - once per virtual interface
 ...
 CpswProxy_close() - once per virtual interface
 CpswProxy_deinit() - once per core, after all virtual interfaces are
                      closed

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>