ethremotecfg: CpswProxyServer: Validate required MAC ports
Since there is a one-to-one mapping between virtual MAC ports and
hardware MAC ports, we have to ensure that the required MAC ports are
actually being enabled.

The list of enabled MAC ports will be passed to CpswProxyServer, which
will check against the virtual ports configured in MAC mode. An error
will be returned if a required MAC port is not being enabled.

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>