ethfw: Remove global ALL_MULTI setting
All-multicast RX filter was applied on CPSW while it was meant only for
port in MAC-only mode.

The unintended effect was that registered and unregistered multicast was
enabled on all VLANs and also on unknown VLAN.

Instead, registered and unregistered multicast will be enabled only on
VLANs where MAC-only ports are present.  At this time, that is, VLAN 0.
There is no new specific action for this, as it was already taken care.

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>