ETHFW-1876 - apps: Intercore memories as 'NOLOAD' type
Linux remoteproc loader is optimized to zero out the delta between
filesz and memsz, where filesz is what gets actually loaded.

Ethernet Firmware was generating two extra segments which had filesz of
zero but memsz of positive value, so remoteproc loader code tried to
zero initialize this area.

Changing the section type to NOLOAD prevents the extra segments to be
generated in the program headers.

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>
Diagnosed-by: Suman Anna <s-anna@ti.com>