utils: board: Update MAC port mapped to QuadETH board
First QuadETH board is connected via Serdes 2 lane 2, which is now being
mapped to Q/SGMII lane 1. The second QuadETH board is connected via
Serdes 2 lane 3, which is mapped to Q/SGMII lane 2.

So the QSGMII sub ports in the first QuadETH board are MAC ports 3, 4
and 5.  And the sub ports in the second QuadETH board will be MAC ports
6, 7 and 8.

Note that the second QuadETH board is not yet supported.

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>