board: j784s4: MAC address pool from EEPROM and static ones
Populate MAC address pool with address read from QSGMII board's EEPROM
and the remaining ones with those from static pool.

This approach ensures that the first MAC address to be used are from
EEPROM (hence unique) and then ones for extended, concurrent usecases
are from static pool (which can lead to duplicated addresses if multiple
devices are on the same network).

Also, a warning will be printed if static addresses are used to
populated the pool.

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>