apps: server: MAC address pool of actual required size
Request MAC address pool to be populated with the actual number of MAC
address required for max concurrency:

 * 1 x MAC address for Ethernet Firmware
 * 2 x MAC address for mpu1_0 virtual switch and MAC-only ports (Linux, 1 for QNX)
 * 2 x MAC address for mcu2_1 virtual switch and MAC-only ports (RTOS)
 * 1 x MAC address for mcu2_1 virtual switch port (AUTOSAR)

So a total of 5 MAC addresses will be populated.  In some cases, the
provisioned addresses in EEPROM will be sufficient, but in others, like
J7200 case, static addresses will have to be used.

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>