ETHFW-2028 - board: Add VSC8514 wait time after NRESET de-assert
As per VSC8514 QSGMII PHY, there is 105 msecs minimum wait time (tWAIT)
required between PHY NRESET de-asserted and access of the SMI interface.
IOW, PHY registers should not be read until after this wait time.

By not meeting this wait time before, EthFw was reading invalid PHY ID
register values and hence binding to the wrong driver (i.e. generic PHY
driver) which caused the interfaces not to be functional.

NRESET gpio line is driven by either EthFw or bootloader.  Bootloader
will not wait for this wait time (as it would unnecesarily impact the
rest of the boot sequence), so EthFw is unconditionally waiting for
tWAIT msecs as long as QSGMII expansion board is used.

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>