ETHFW-2152 - ethfw: Calculate RM needs based on client cores
Previously, the resources (TX chan, RX flow, MAC addr) needed in a given
ETHFW configuration were computed based on the number of clients.
That approach double counter the resources needed for mutually
exclusive clients like RTOS and AUTOSAR virtual clients running on
mcu2_1.

Going forward, the resources are computed based on the max number of
clients supported for each core.  For instance, again mcu2_1:
 - RTOS: 1 x virtual switch port + 1 x virtual MAC port
 - MCAL: 1 x virtual switch port

So, the total resources to be allocated for mcu2_1 will be for up to
2 virtual clients.

This ensures that a single ETHFW configuration can continue to be used
in SDK to offer virtual Ethernet functionality to a given core which
can run clients of different OSes.

Signed-off-by: Doredla Sudheer Kumar <s-doredla@ti.com>