ETHFW-1964 - apps: server: ETHFW boot time profiling
Added initial support for ETHFW boot time profiling. The following boot
milestones are profiled:

1. IPC server ready
2. Ethernet port got link up
3. Host port is ready
4. TCP/IP stack got IP
5. PTP stack is initialized

The intended setup for profiling is one that decouples PHY link up time
as it depends on PHY itself, auto-negotiation or fixed mode,
boot-strapping, etc.  So a MAC-to-MAC connection using SGMII bridge is
used for profiling.

UART prints are disabled for the active phase of the profiling - when
the timestamps are being captured, and re-enabled later for convinience
to print the boot profiling results.

At build time, user must append ETHFW_BOOT_TIME_PROFILING=yes to the
make command for profiling to be enabled on ETHFW server image.  It's
worth noting that profiling is enabled only in CCS image for simplicity.

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>