ETHFW-2225 - Ethfw: VEPA support for J784S4
Key features in this patch:

1. Added Vepa Utils lib for vepa table
2. Seperate packet duplication flow for broadcast and shared multicast packets
3. Send duplicated packets to respective virtual switch ports based on private VLAN
4. Added policer table partition based on priority by the application

VEPA enables to ingress and egress packet on the same host port. This is done
by enabling multihost bit in ale_ctr2 control register. This has
enabled us to forwared the packets to clients via host port and to not use
shared memory netifs. This will increase the intercore packet transmission
performance.

As packet duplication (for packet with more than 1 recepient) is not done by cpsw
hardware in host port right now, ethfw is doing the same job and sending to the
intended clients. This is done by putting a private vlan id to the packet and
forwarding based on the private vlan.

VEPA will be enabled by default for j784s4. Please note that PROXY_ARP_HANDLING
and ETHAPP_ENABLE_INTERCORE_ETH will be disabled as clients are receiving
the packets directly.

Signed-off-by: Vaibhav Jindal <v-jindal@ti.com>