ethremotecfg: server: Workaround for stats not reset during recovery
While CPSW statistics counters are being cleared during CPSW module
reset, the software counters maintained by Enet LLD are not.  This is a
problem for the error detection mechanism implemented in ETHFW as it
assumes that statistics are cleared.

As a workaround, the starting point for error detection is not a stats
counter of 0, but rather the last saved value during previous recovery.

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>