ethremotecfg: server: Misc fixes for shared multicast untagged traffic
For VLAN untagged traffic, the ALE table and policer table entries have
to be created using the actual VLAN used in hardware, that is, the
default VLAN id for switch ports.  For the local tables in software,
untagged traffic is being added as VLAN id 0.  For VLAN tagged traffic,
no special handling required.

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>
Signed-off-by: Vaibhav Jindal <v-jindal@ti.com>