internal_docs: User guide documentation for VEPA
We are showing the new topology diagram with new data path flow with VEPA. We are
handling the unicast packets between cores directly without R5 (EthFw )to intervene.
Multicast and  broadcast packets are handled via a secondary RX flow on which EthFw
duplicates the packet, add a private vlan tag and sends the packet back on host port
so that it can be redirected to respective virtual switch port based on policer.

Signed-off-by: Vaibhav Jindal <v-jindal@ti.com>