ethremotecfg: server: ethfw_api: Workaround for promiscuous mode issue on MAC-only port
We achieve promiscuous mode on MAC-only port by enabling copy all frames (caf) bit in ALE.
Issue is that VLAN untagging doesn't happen (irrespective of ingressed packet type)
in hardware if caf bit is set. Due to this limitation packets are coming with MAC-onlpy port's
default VLAN tagged in promiscusous mode in linux client which is not expected behaviour. As a
workaround, we have made default VLAN to be 0 for MAC-only ports. Hence even when packets are
received with VLAN 0 in Linux client it considers it as no VLAN tagged.

Fixes: ETHFW-2536

Signed-off-by: Vaibhav Jindal <v-jindal@ti.com>