apps: tap: Add support for bridge interface creation
Add the shell script and the corresponding service to run the shell script which creates the bridge interface as the master of the TAP interface and the Virtual Switch Interface in Linux.

The bridge_setup.service can be enabled similar to launch_tap.service following which the bridge interface shall automatically be created.

The bridge interface is required when both the TAP interface and the Virtual Switch Interface are in the same subnet which corresponds to DHCP based IP assignment. The bridge interface will not be required if both the TAP interface and the Virtual Switch Interface are in different subnets which shall be the case when using Static IP.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>