ethremotecfg: server: Default VLAN of all ports set to VLAN_ID_MAX at initialization, as 0U is reserved for MAC only ports
We achieve promiscuous mode on MAC-only port by enabling copy all frames (caf) bit in ALE.
VLAN untagging doesn't happen in hardware if caf bit is set causing packets to come with
default VLAN tagged in promiscusous mode in linux client (not expected behaviour). As a
workaround, we have made default VLAN to be 0 for MAC-only ports.

To support this default VLAN of all ports is set to VLAN_ID_MAX at initialization by the application.
Issue is that if we kept default VLAN to be 0 then CpswAle_addDefaultEntries will
overwrite existing ALE entry for VLAN 0 created for MAC-only port (causing wrong ALE entry to be added).
This issue only comes on J784S4 because we are not setting all ports' hardware configuration
parameters in ethfw application file (main.c) which is absolutely fine, as application only needs
to configure the ports which it requires.

In short this entry -> Vlanid: 0000, UTagged: 0, Mult: 1ff, UMult: 1ff, Member: 1ff
holds wrong information in ALE causing problem.

Fixes: ETHFW-2536

Signed-off-by: Vaibhav Jindal <v-jindal@ti.com>