proxy-server: moved clientObj into coreObj in server implementation
In the heirarchy of handling clients, Server Object holds the info
of the remote cores in coreObj which in turn contains one ore more
clientObj. In the current implementation each remote core supports
two remote clients (virtual port) support. If in future this support
needs to be extended to more number simply increase
CPSWPROXYSERVER_REMOTE_CLIENT_MAX, which will be reflected to all remote
cores.

Signed-off-by: Meghana Malladi <m-malladi@ti.com>