ethremotecfg: server: mcast: Enable multicast (shared and exclusive) packets from clients
When client tries to send multicast packets to external ports there we are seeing packet drop on host port.
This is happening because there is no entry for multicast address with host port vlan in ALE.
As at the time of registering multicast traffic (via ETHREMOTECFG_CMD_ADD_FILTER_MAC) we are adding ALE entry for switch ports only.

Fix is to add an ALE entry for host port VLAN with the multicast address at the time of registration.

Fixes - ETHFW-2576

Signed-off-by: Vaibhav Jindal <v-jindal@ti.com>