ethremotecfg: protocol: Removal of EthRemoteCfg_RxFlowInfo from protocol header file
EthRemoteCfg_RxFlowInfo is required for creating custom policers for rx flows for QoS.
But putting it in protocol header file creates it a dependency for clients as well which is not required.
As only ethfw needs to have this information with it, we have moved it to common header file for ethfw.

Signed-off-by: Vaibhav Jindal <v-jindal@ti.com>