EthFw UT: server: Fix for port mirroring UT
Calling EthFwPortMirroring_init with NULL causing test case to fail.
Every parameter in EthFw_Config structure which is a pointer need to
be initialized in setUp so that the values are being passed to EthFw_init.
Without this values at the address of these pointers will be garbage or invalid

Fixes: ETHFW-2933

Signed-off-by: Vaibhav Jindal <v-jindal@ti.com>