ethfw: proxy_server and UT: Fixes for double attach and related UT fixes.
- Fixed double attach issue of giving false positive with
just and warning print. If a client tries to attach again
for the same virtual port, throw and error.

- Corresponding to above fix, entire UT infra needs to
be update to attach and detach per test, also proxy
object is private  to cpsw_proxy.c so have to use
exposed proxy apis for attaching instead of using
direct remotecfg CMD.

- Multiple detaches from above fix causes and assert
on server due to incorrect release of MCM handles in
detach handling, fixed that.

- EthFwUT_attachCmdNegTest test is invalid, since any
virtport will still pass through by server, and virtport is
the only param used to attach. Existing test was using
request and response params to oeprate the test but this
usually happen at a later stage by which attach is already
succesful. Removed the invalid test.

Fixes: ETHFW-2667, ETHFW-2712, ETHFW-3003, ETHFW-3004

Signed-off-by: Sonu Alam <s-alam@ti.com>