docs: ethfw_mcast: ETHFW exclusive multicast address on different VLANs
In case when the same client registers M in a a different VLAN,
we are allowing the same client to register M with different VLANs.
It is worth noting that M can be registered with different VLAN V1 and V2
by same client C1 only, if another client C2 tries to register M with V3
(totally different VLAN) then that request will result an error.

Signed-off-by: Vaibhav Jindal <v-jindal@ti.com>