ethfw: mts: Add MTS demo test for sync accuracy.
- updated client coupler to calculate the accuracy between
local sync time and remote CPTS time. Idea behind is when
a timer overflow happens, isr gets called and at the same time
PWM triggers a HW push event. So calculate the local sync time
based on rate and offset calculation and compare it with CPTS
time receied and print the delta.

Fixes: ETHFW-2203, ETHFW-2211

Signed-off-by: Sonu Alam <s-alam@ti.com>